Added category view on main page. Each category has a thread counter.
Added page to view posts per category. Users with role Admin can delete the thread, including all answers in it.

Added details page to view details for each thread.
Added button to delete answer. This button is shown only the users with Admin role and to the Author of the answer.
Added button to edit answer. This button is shown only the users with Admin role and to the Author of the answer.

By clicking on the poster name, you can see all threads, that he has posted.

+++ Forum/Forum/Controllers/CategoryController.cs
@@ -131,6 +131,16 @@ namespace Forum.Controllers
             {
                 return new HttpStatusCodeResult(HttpStatusCode.NotFound);
             }
+
+            var isEmpty = db.Articles.Where(c => c.CategoryId == id).Count();
+
+            if(isEmpty > 0)
+            {
+                ModelState.AddModelError(string.Empty, "You can not delete a category which contains threads!");
+
+                return View(category);
+            }
+
             db.Categories.Remove(category);
             db.SaveChanges();
